+++ NEWS.md
@@ -1,3 +1,5 @@
 # tidycovid19 0.0.0.9000
 
-* initial in-development version on Github
+* `shiny_covid19_spread()`: Added option to copy plot code to clipboard
+
+* Initial in-development version on Github